requirements.txt
bson==0.5.1
pydantic==2.4.2
python-telegram-bot==20.6
pymongo==4.5.0
python-dateutil==2.8.2